converting old system to new

+++ JamChester/Assets/Project Files/Scripts/TrophyStand/TrophyShelf.cs
@@ -14,7 +14,7 @@ public class TrophyShelf : MonoBehaviour {
 	// Use this for initialization
 	void Start () {
         for(int i = 0; i < slots.Length; i++) {
-            slots[i] = Instantiate(possibleSlots[(int)Random.Range(0, possibleSlots.Length)], transform.position + slotSpacing*i,Quaternion.identity) as TrophySlot;
+            slots[i] = Instantiate(possibleSlots[Random.Range(0, possibleSlots.Length)], transform.position + slotSpacing*i,Quaternion.identity) as TrophySlot;
             slots[i].owner = gameObject.GetComponent<TrophyShelf>();
             slots[i].gameObject.SetActive(true);
         }
